Move everything from config to configparser
