+++ config/configparser/testdata/duplicate-exporter.yaml
