+++ config/configparser/testdata/duplicate-extension.yaml
