+++ config/configparser/testdata/duplicate-pipeline.yaml
