+++ config/configparser/testdata/duplicate-processor.yaml
