+++ config/configparser/testdata/duplicate-receiver.yaml
