+++ config/configparser/testdata/invalid-exporter-name-after-slash.yaml
