+++ config/configparser/testdata/invalid-exporter-section.yaml
