+++ config/configparser/testdata/invalid-exporter-type.yaml
