+++ config/configparser/testdata/invalid-extension-name-after-slash.yaml
