+++ config/configparser/testdata/invalid-extension-type.yaml
