+++ config/configparser/testdata/invalid-pipeline-name-after-slash.yaml
