+++ config/configparser/testdata/invalid-pipeline-section.yaml
