+++ config/configparser/testdata/invalid-pipeline-type.yaml
