+++ config/configparser/testdata/invalid-processor-name-after-slash.yaml
