+++ config/configparser/testdata/invalid-processor-type.yaml
