+++ config/configparser/testdata/invalid-receiver-name-after-slash.yaml
