+++ config/configparser/testdata/invalid-receiver-section.yaml
