+++ config/configparser/testdata/invalid-receiver-sub-config.yaml
