+++ config/configparser/testdata/invalid-receiver-type.yaml
