+++ config/configparser/testdata/invalid-sequence-value.yaml
