+++ config/configparser/testdata/invalid-service-extensions-section.yaml
