+++ config/configparser/testdata/unknown-pipeline-type.yaml
