+++ config/configparser/testdata/unknown-receiver-type.yaml
