+++ receiver/hostmetricsreceiver/factory.go
@@ -23,8 +23,8 @@ import (
 	"go.uber.org/zap"
 
 	"go.opentelemetry.io/collector/component"
-	"go.opentelemetry.io/collector/config"
 	"go.opentelemetry.io/collector/config/configmodels"
+	"go.opentelemetry.io/collector/config/configparser"
 	"go.opentelemetry.io/collector/consumer"
 	"go.opentelemetry.io/collector/receiver/hostmetricsreceiver/internal"
 	"go.opentelemetry.io/collector/receiver/hostmetricsreceiver/internal/scraper/cpuscraper"
@@ -93,7 +93,7 @@ func customUnmarshaler(componentViperSection *viper.Viper, intoCfg interface{}) 
 
 	cfg.Scrapers = map[string]internal.Config{}
 
-	scrapersViperSection, err := config.ViperSubExact(componentViperSection, scrapersKey)
+	scrapersViperSection, err := configparser.ViperSubExact(componentViperSection, scrapersKey)
 	if err != nil {
 		return err
 	}
@@ -108,7 +108,7 @@ func customUnmarshaler(componentViperSection *viper.Viper, intoCfg interface{}) 
 		}
 
 		collectorCfg := factory.CreateDefaultConfig()
-		collectorViperSection, err := config.ViperSubExact(scrapersViperSection, key)
+		collectorViperSection, err := configparser.ViperSubExact(scrapersViperSection, key)
 		if err != nil {
 			return err
 		}
